+++ hkl/tests/test_configuration.py
@@ -1,10 +1,24 @@
+import pathlib
 from contextlib import nullcontext as does_not_raise
+from dataclasses import MISSING
 
 import pytest
+from apischema import ValidationError
+from apischema import deserialize
 
 from .. import DiffractometerConfiguration
 from ..configuration import EXPORT_FORMATS
-from ..configuration import REQUIRED_CONFIGURATION_KEYS_TYPES
+from ..configuration import DCConfiguration
+from ..configuration import DCConstraint
+from ..configuration import DCLattice
+from ..configuration import DCReflection
+from ..configuration import DCSample
+from ..util import Constraint
+from ..util import new_lattice
+from . import TWO_PI
+
+
+TEST_CONFIG_FILE = "data/e4c-config.json"
 
 
 def test_e4cv(e4cv):
@@ -59,35 +73,289 @@ def test_format(fmt, e4cv):
         config.restore(cfg)  # test restore with automatic type recognition
 
 
+@pytest.mark.parametrize("file", [None, TEST_CONFIG_FILE])  # default or restored config
 @pytest.mark.parametrize("action", "rm set".split())  # remove or set keys incorrectly
 @pytest.mark.parametrize(
-    "key, value, failure",
-    [[k, object, AssertionError] for k in REQUIRED_CONFIGURATION_KEYS_TYPES]
+    "key, value, failure",  # fmt: off
+    [
+        [k, object, (KeyError, TypeError, ValidationError, ValueError)]
+        for k in DCConfiguration.__dataclass_fields__
+    ],  # fmt: off
 )
-def test_validation_fails(action, key, value, failure, tardis):
-    assert len(tardis.calc._samples) == 1
-    assert tardis.calc.sample.name == "main"
+def test_validation_fails(file, action, key, value, failure, e4cv):
+    agent = DiffractometerConfiguration(e4cv)
+    assert isinstance(agent, DiffractometerConfiguration), f"{agent}"
 
-    with pytest.raises(TypeError):
-        cfg = DiffractometerConfiguration("wrong diffractometer object")
-        cfg.validate_config_dict({})
+    if file is not None:
+        path = pathlib.Path(__file__).parent / file
+        assert path.exists(), f"{path}"
+        with open(path) as f:
+            agent.restore(f.read())
 
-    with pytest.raises(TypeError):
-        cfg = DiffractometerConfiguration(tardis)
-        cfg.validate_config_dict("wrong configuration object")
+    data = agent.export("dict")
+    assert isinstance(data, dict), f"{type(data)=}"
 
-    with pytest.raises(failure):
-        cfg = DiffractometerConfiguration(tardis)
-        assert isinstance(cfg, dict), f"{cfg}"
+    if action == "rm":
+        data.pop(key)
+        # Determine if key is optional (default/factory is defined).  Empirical.
+        attr = DCConfiguration.__dataclass_fields__[key]
+        if attr.default_factory != MISSING or attr.default != MISSING:
+            failure = None  # OK if not provided
+    elif action == "set":
+        data[key] = value
 
-        if action == "rm":
-            cfg.pop(key)
-        elif action == "set":
-            cfg[key] = value
-        cfg.validate_config_dict(cfg)
+    if failure is None:
+        agent.restore(data)
+    else:
+        with pytest.raises(failure):
+            agent.restore(data)
 
 
-# TODO: test sample dictionary
-# TODO: test reflections dictionary
-# TODO: test that diffractometer is updated by config.restore()
-# TODO: Test the `clear` flag for config.restore()
+def common_DC_dataclass_tests(dc_class, data, key, value, failure, val_arg):
+    """Code used to test the DCxyz classes."""
+    agent = deserialize(dc_class, data)
+    assert isinstance(agent, dc_class)
+    assert key in data
+
+    data.pop(key)
+    with pytest.raises(ValidationError):
+        deserialize(dc_class, data)
+
+    data[key] = value
+    with failure:
+        agent = deserialize(dc_class, data)
+        assert isinstance(agent, dc_class)
+        agent.validate(val_arg)
+
+
+@pytest.mark.parametrize("key", "low_limit high_limit value".split())
+# fit is a boolean, existing validation testing is sufficient for now
+@pytest.mark.parametrize(
+    "value, failure",
+    [
+        [-360.01, pytest.raises(ValueError)],
+        [-360, does_not_raise()],
+        [0, does_not_raise()],
+        [360, does_not_raise()],
+        [360.01, pytest.raises(ValueError)],
+        ["0", pytest.raises(ValidationError)],
+    ],
+)
+def test_DCConstraint_fails(key, value, failure):
+    # all attributes are required
+    data = {
+        "low_limit": 0.0,
+        "high_limit": 0.0,
+        "value": 0.0,
+        "fit": True,
+    }
+    common_DC_dataclass_tests(DCConstraint, data, key, value, failure, f"testing DCConstraint.{key=}")
+
+
+@pytest.mark.parametrize("key", "a b c alpha beta gamma".split())
+@pytest.mark.parametrize(
+    "value, failure",
+    [
+        [-1, pytest.raises(ValueError)],
+        [0, pytest.raises(ValueError)],
+        [1e-7, pytest.raises(ValueError)],
+        [1, does_not_raise()],
+        [179.99, does_not_raise()],
+        [180, does_not_raise()],
+        [10_000, does_not_raise()],
+        [100_000_000, pytest.raises(ValueError)],
+    ],
+)
+def test_DCLattice_fails(key, value, failure, e4cv):
+    data = {
+        "a": 4,
+        "b": 5,
+        "c": 6,
+        "alpha": 8,
+        "beta": 9,
+        "gamma": 10,
+    }
+    if key in "alpha beta gamma".split() and value > 180 - 1e-6:
+        failure = pytest.raises(ValueError)
+
+    agent = DiffractometerConfiguration(e4cv)
+    common_DC_dataclass_tests(DCLattice, data, key, value, failure, agent)
+
+
+@pytest.mark.parametrize(
+    "key, value, failure",
+    [
+        ["wavelength", -1, pytest.raises(ValueError)],
+        ["wavelength", 0, pytest.raises(ValueError)],
+        ["wavelength", 0.01, does_not_raise()],
+        ["wavelength", 500, does_not_raise()],
+        ["wavelength", 1_000_000.0, does_not_raise()],
+        ["wavelength", 100_000_000, pytest.raises(ValueError)],
+        ["reflection", {"h": 12.4, "k": 0, "l": 0}, does_not_raise()],
+        ["reflection", {"h": 0, "k": 12.4, "l": 0}, does_not_raise()],
+        ["reflection", {"h": 0, "k": 0, "l": 12.4}, does_not_raise()],
+        ["reflection", {"h": -12.4, "k": 0, "l": 0}, does_not_raise()],
+        ["reflection", {"h": 0, "k": -12.4, "l": 0}, does_not_raise()],
+        ["reflection", {"h": 0, "k": 0, "l": -12.4}, does_not_raise()],
+        ["reflection", {"h": 20_000, "k": 0, "l": 0}, pytest.raises(ValueError)],
+        ["reflection", {"h": -20_000, "k": 0, "l": 0}, pytest.raises(ValueError)],
+        ["reflection", {"h": 0, "k": 20_000, "l": 0}, pytest.raises(ValueError)],
+        ["reflection", {"h": 0, "k": -20_000, "l": 0}, pytest.raises(ValueError)],
+        ["reflection", {"h": 0, "k": 0, "l": 20_000}, pytest.raises(ValueError)],
+        ["reflection", {"h": 0, "k": 0, "l": -20_000}, pytest.raises(ValueError)],
+        ["position", {"omega": 360, "chi": 0, "phi": 0, "tth": 0}, does_not_raise()],
+        ["position", {"omega": 0, "chi": 360, "phi": 0, "tth": 0}, does_not_raise()],
+        ["position", {"omega": 0, "chi": 0, "phi": 360, "tth": 0}, does_not_raise()],
+        ["position", {"omega": 0, "chi": 0, "phi": 0, "tth": 360}, does_not_raise()],
+        ["position", {"omega": -360, "chi": 0, "phi": 0, "tth": 0}, does_not_raise()],
+        ["position", {"omega": 0, "chi": -360, "phi": 0, "tth": 0}, does_not_raise()],
+        ["position", {"omega": 0, "chi": 0, "phi": -360, "tth": 0}, does_not_raise()],
+        ["position", {"omega": 0, "chi": 0, "phi": 0, "tth": -360}, does_not_raise()],
+        ["position", {"omega": 360.01, "chi": 0, "phi": 0, "tth": 0}, pytest.raises(ValueError)],
+        ["position", {"omega": 0, "chi": 360.01, "phi": 0, "tth": 0}, pytest.raises(ValueError)],
+        ["position", {"omega": 0, "chi": 0, "phi": 360.01, "tth": 0}, pytest.raises(ValueError)],
+        ["position", {"omega": 0, "chi": 0, "phi": 0, "tth": 360.01}, pytest.raises(ValueError)],
+        ["position", {"omega": -360.01, "chi": 0, "phi": 0, "tth": 0}, pytest.raises(ValueError)],
+        ["position", {"omega": 0, "chi": -360.01, "phi": 0, "tth": 0}, pytest.raises(ValueError)],
+        ["position", {"omega": 0, "chi": 0, "phi": -360.01, "tth": 0}, pytest.raises(ValueError)],
+        ["position", {"omega": 0, "chi": 0, "phi": 0, "tth": -360.01}, pytest.raises(ValueError)],
+    ],
+)
+def test_DCReflection_fails(key, value, failure, e4cv):
+    data = {
+        "reflection": {"h": 0, "k": 0, "l": 0},
+        "position": {"omega": 0, "chi": 0, "phi": 0, "tth": 0},
+        "wavelength": 1,
+        "orientation_reflection": True,
+    }
+    agent = DiffractometerConfiguration(e4cv)
+    common_DC_dataclass_tests(DCReflection, data, key, value, failure, agent)
+
+
+@pytest.mark.parametrize(
+    "key, value, failure",
+    [
+        ["name", "", pytest.raises(ValueError)],
+        ["name", "   ", pytest.raises(ValueError)],
+        ["lattice", object, pytest.raises(TypeError)],
+        ["lattice", {}, pytest.raises(ValidationError)],
+        ["lattice", {"error", object}, pytest.raises(TypeError)],
+        ["reflections", [object], pytest.raises(TypeError)],
+        ["U", [object], pytest.raises(TypeError)],
+        ["UB", [object], pytest.raises(TypeError)],
+        ["U", [[1, 0, object], [0, 0, 0], [0, 0, 0]], pytest.raises(TypeError)],
+        ["UB", [[1, 0, object], [0, 0, 0], [0, 0, 0]], pytest.raises(TypeError)],
+        ["U", [[1, 0, 0], [0, 0, 0]], pytest.raises(ValueError)],
+        ["UB", [[1, 0, 0], [0, 0, 0]], pytest.raises(ValueError)],
+    ],
+)
+def test_DCSample_fails(key, value, failure, e4cv):
+    data = {
+        "name": "vibranium",
+        "lattice": {"a": 4, "b": 4, "c": 4, "alpha": 90, "beta": 90, "gamma": 90},
+        "reflections": [],
+        "U": [[1.0, 0.0, 0.0], [0.0, 1.0, 1.0], [0.0, 0.0, 1.0]],
+        "UB": [[TWO_PI, 0.0, 0.0], [0.0, TWO_PI, 0.0], [0.0, 0.0, TWO_PI]],
+    }
+    agent = DiffractometerConfiguration(e4cv)
+    common_DC_dataclass_tests(DCSample, data, key, value, failure, agent)
+
+
+@pytest.mark.parametrize("clear", [True, False, object])
+def test_diffractometer_restored(clear, e4cv):
+    # -------------------------------- default configuration
+    mode_before = e4cv.engine.mode
+    positions_before = e4cv.RealPosition
+    reciprocal_positions_before = e4cv.PseudoPosition
+    constraints_before = e4cv._constraints_for_databroker
+
+    assert round(e4cv.calc.wavelength, 2) == 1.55
+    assert e4cv.engine.mode == mode_before
+    assert len(e4cv.calc._samples) == 1
+    assert e4cv.calc.sample.name == "main"
+    assert len(e4cv.calc.sample.reflections_details) == 0
+
+    # -------------------------------- perturb from the defaults
+    mode_changed = e4cv.engine.modes[-1]
+    constraints_test = {
+        "omega": Constraint(-10, 80, 10, 1),
+        "chi": Constraint(-10, 91, 45, 0),
+        "phi": Constraint(-10, 91, 90, 0),
+        "tth": Constraint(-10, 120, 20, 1),
+    }
+    wavelength_test = 0.25
+
+    e4cv.engine.mode = mode_changed
+    assert e4cv.engine.mode == mode_changed
+
+    assert e4cv._constraints_for_databroker == constraints_before
+    assert e4cv.calc.wavelength != wavelength_test
+
+    e4cv.apply_constraints(constraints_test)
+    constraints_changed = e4cv._constraints_for_databroker
+    assert constraints_changed != constraints_before
+
+    e4cv.calc.wavelength = wavelength_test
+    assert e4cv.calc.wavelength == wavelength_test
+
+    # modify the sample lattice
+    orthorhombic = new_lattice(4, 5, 6)
+    e4cv.calc.sample.lattice = orthorhombic
+    assert e4cv.calc.sample.lattice == orthorhombic
+
+    # add a reflection
+    assert len(e4cv.calc.sample.reflections_details) == 0
+    e4cv.calc.sample.add_reflection(0.1, 0, 0, (1, 2, 3, 4))
+    assert len(e4cv.calc.sample.reflections_details) == 1
+
+    # -------------------------------- restore config from file
+    test_file = pathlib.Path(__file__).parent / TEST_CONFIG_FILE
+    assert test_file.exists()
+
+    agent = DiffractometerConfiguration(e4cv)
+    context = does_not_raise() if isinstance(clear, bool) else pytest.raises(TypeError)
+    with context:
+        full_config_before = agent.export("dict")
+        with open(test_file) as f:
+            agent.restore(f.read(), clear=clear)
+
+    # -------------------------------- these should not have been changed
+    assert e4cv.calc.wavelength == wavelength_test, "wavelength changed"
+    assert e4cv.RealPosition == positions_before, "RealPosition changed"
+    assert e4cv.PseudoPosition == reciprocal_positions_before, "PseudoPosition changed"
+
+    # -------------------------------- changes conditional on clear
+
+    full_config_after = agent.export("dict")
+    full_config_before.pop("datetime")
+    full_config_after.pop("datetime")
+
+    if isinstance(clear, bool):
+        assert full_config_after != full_config_before
+        assert e4cv._constraints_for_databroker == constraints_before
+        assert e4cv._constraints_for_databroker != constraints_changed
+        # one additional sample
+        assert len(e4cv.calc._samples) == 2
+        assert list(e4cv.calc._samples) == "main vibranium".split()
+
+        if clear:
+            assert e4cv.engine.mode == mode_before
+            assert e4cv.engine.mode != mode_changed
+            assert e4cv.calc._samples["main"].lattice != orthorhombic
+            assert len(e4cv.calc._samples["main"].reflections_details) == 0
+        else:
+            # switch samples
+            e4cv.calc.sample = "vibranium"
+            assert len(e4cv.calc.sample.reflections_details) == 3
+            e4cv.calc.sample.lattice.a == TWO_PI
+            e4cv.calc.sample.lattice.b == TWO_PI
+            e4cv.calc.sample.lattice.c == TWO_PI
+    else:
+        # when restore() failed
+        assert full_config_after == full_config_before
+        assert len(e4cv.calc._samples) == 1
+        assert list(e4cv.calc._samples) == "main".split()
+        assert e4cv.engine.mode != mode_before
+        assert e4cv.engine.mode == mode_changed
+        assert e4cv.calc._samples["main"].lattice == orthorhombic
+        assert len(e4cv.calc._samples["main"].reflections_details) == 1